requirements.txt
attrs==24.3.0
autopep8==2.3.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
Flask==3.1.0
flatbuffers==24.12.23
humanfriendly==10.0
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.2
pycodestyle==2.12.1
PyMatting==1.1.13
referencing==0.35.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.14.1
sympy==1.13.3
tifffile==2024.12.12
tqdm==4.67.1
urllib3==2.3.0
Werkzeug==3.1.3
